test: factor boot-tools out of test-postalSvc
 - wrapZoe, makeZoeWithBundles
